refactor: Cached* -> Fs*
